+++ test/unit/grant-types/authorization-code-grant-type_test.js
@@ -72,17 +72,17 @@ describe('AuthorizationCodeGrantType', function() {
       };
       const handler = new AuthorizationCodeGrantType({ accessTokenLifetime: 120, model: model });
 
-      sinon.stub(handler, 'validateScope').returns('foobiz');
+      sinon.stub(handler, 'validateScope').returns(['foobiz']);
       sinon.stub(handler, 'generateAccessToken').returns(Promise.resolve('foo'));
       sinon.stub(handler, 'generateRefreshToken').returns(Promise.resolve('bar'));
       sinon.stub(handler, 'getAccessTokenExpiresAt').returns(Promise.resolve('biz'));
       sinon.stub(handler, 'getRefreshTokenExpiresAt').returns(Promise.resolve('baz'));
 
-      return handler.saveToken(user, client, 'foobar', 'foobiz')
+      return handler.saveToken(user, client, 'foobar', ['foobiz'])
         .then(function() {
           model.saveToken.callCount.should.equal(1);
           model.saveToken.firstCall.args.should.have.length(3);
-          model.saveToken.firstCall.args[0].should.eql({ accessToken: 'foo', authorizationCode: 'foobar', accessTokenExpiresAt: 'biz', refreshToken: 'bar', refreshTokenExpiresAt: 'baz', scope: 'foobiz' });
+          model.saveToken.firstCall.args[0].should.eql({ accessToken: 'foo', authorizationCode: 'foobar', accessTokenExpiresAt: 'biz', refreshToken: 'bar', refreshTokenExpiresAt: 'baz', scope: ['foobiz'] });
           model.saveToken.firstCall.args[1].should.equal(client);
           model.saveToken.firstCall.args[2].should.equal(user);
           model.saveToken.firstCall.thisValue.should.equal(model);