+++ test/unit/grant-types/password-grant-type_test.js
@@ -45,17 +45,17 @@ describe('PasswordGrantType', function() {
       };
       const handler = new PasswordGrantType({ accessTokenLifetime: 120, model: model });
 
-      sinon.stub(handler, 'validateScope').returns('foobar');
+      sinon.stub(handler, 'validateScope').returns(['foobar']);
       sinon.stub(handler, 'generateAccessToken').returns('foo');
       sinon.stub(handler, 'generateRefreshToken').returns('bar');
       sinon.stub(handler, 'getAccessTokenExpiresAt').returns('biz');
       sinon.stub(handler, 'getRefreshTokenExpiresAt').returns('baz');
 
-      return handler.saveToken(user, client, 'foobar')
+      return handler.saveToken(user, client, ['foobar'])
         .then(function() {
           model.saveToken.callCount.should.equal(1);
           model.saveToken.firstCall.args.should.have.length(3);
-          model.saveToken.firstCall.args[0].should.eql({ accessToken: 'foo', accessTokenExpiresAt: 'biz', refreshToken: 'bar', refreshTokenExpiresAt: 'baz', scope: 'foobar' });
+          model.saveToken.firstCall.args[0].should.eql({ accessToken: 'foo', accessTokenExpiresAt: 'biz', refreshToken: 'bar', refreshTokenExpiresAt: 'baz', scope: ['foobar'] });
           model.saveToken.firstCall.args[1].should.equal(client);
           model.saveToken.firstCall.args[2].should.equal(user);
           model.saveToken.firstCall.thisValue.should.equal(model);